Fix types in test suite.

+++ examples/random.rs
@@ -1,4 +1,5 @@
 // use rand::Rng;
+use num_bigint::BigInt;
 use num_rational::BigRational;
 use rand::distributions::Standard;
 use rand::Rng;
@@ -10,9 +11,11 @@ fn main() {
   dbg!(random_between(10, 100, &mut rng)
     .iter()
     .fold(0, |a, b| a + b));
-  dbg!(random_between(10, 100, &mut rng).iter().sum::<usize>());
+  dbg!(random_between(10, 100, &mut rng).iter().sum::<BigInt>());
   dbg!(random_between_zero(10, 100, &mut rng));
-  dbg!(random_between_zero(10, 100, &mut rng).iter().sum::<isize>());
+  dbg!(random_between_zero(10, 100, &mut rng)
+    .iter()
+    .sum::<BigInt>());
 
   // dbg!(random(10, 1000, &mut rng));
   let val: ConvexPolygon<BigRational> = rng.sample(Standard);
